refactor(cli): share generator import path helper
